Agrega ruff, script de verificacion y workflow de CI
Backlog #14. ruff.toml con un set corto de reglas, requirements-dev.txt
(ruff y nada mas, para que el runtime siga en dos dependencias),
scripts/check.ps1 con lint + tests + build, y .github/workflows/ci.yml
listo para cuando el repo tenga remoto.

De los 36 hallazgos iniciales, 11 eran RUF012 sobre el `class Meta:
ordering = [...]` de Django: son listas por definicion del framework, no
estado compartido peligroso, asi que la regla se ignora con el motivo
escrito en la config. Los otros 25 se arreglaron:

- 12 lineas largas.
- 4 zip() sin strict=, que ahora revienta si las longitudes dejan de
  coincidir en vez de truncar en silencio.
- una variable `l` que se confunde con un 1 en casi cualquier tipografia.
- un int(round(...)) redundante y dos bloques de imports desordenados.

ruff format no entra: reformatearia 23 archivos para pelear con un estilo
que ya es consistente. El linter busca errores, el formateador impone
gustos, y eso queda escrito en el script.

El autofix borro los `# noqa: BLE001 - se registra y se sigue` junto con
su explicacion. El except amplio sigue siendo deliberado, asi que la razon
volvio como comentario normal.

check.ps1 usa ErrorActionPreference Continue y mira LASTEXITCODE: Django y
npm escriben su salida normal en stderr, y con Stop PowerShell la toma
como error y corta en la primera linea.

291 tests, ruff limpio, build ok.

+++ backend/tests/factories.py
@@ -154,7 +154,7 @@ def build_dataset(
     round_counter = 0
     for i in range(matches):
         map_name = maps[i % len(maps)]
-        won_rounds = int(round(rounds_per_match * my_winrate))
+        won_rounds = round(rounds_per_match * my_winrate)
         match = make_match(
             map_name=map_name,
             index=i,
